+++ cmd/vega/main.go
@@ -16,6 +16,7 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"runtime/debug"
 
 	"github.com/jessevdk/go-flags"
 
@@ -27,11 +28,8 @@ import (
 )
 
 var (
-	// CLIVersionHash specifies the git commit used to build the application. See VERSION_HASH in Makefile for details.
 	CLIVersionHash = ""
-
-	// CLIVersion specifies the version used to build the application. See VERSION in Makefile for details.
-	CLIVersion = ""
+	CLIVersion     = "v0.53.0+dev"
 )
 
 // Subcommand is the signature of a sub command that can be registered.
@@ -48,6 +46,7 @@ func Register(ctx context.Context, parser *flags.Parser, cmds ...Subcommand) err
 }
 
 func main() {
+	setCommitHash()
 	ctx := context.Background()
 	if err := Main(ctx); err != nil {
 		os.Exit(-1)
@@ -77,7 +76,6 @@ func Main(ctx context.Context) error {
 		Wallet,
 		Watch,
 		Tm,
-		Checkpoint,
 		Query,
 		Bridge,
 		paths.Paths,
@@ -94,4 +92,21 @@ func Main(ctx context.Context) error {
 		return err
 	}
 	return nil
+}
+
+func setCommitHash() {
+	info, _ := debug.ReadBuildInfo()
+	modified := false
+
+	for _, v := range info.Settings {
+		if v.Key == "vcs.revision" {
+			CLIVersionHash = v.Value
+		}
+		if v.Key == "vcs.modified" {
+			modified = true
+		}
+	}
+	if modified {
+		CLIVersionHash += "-modified"
+	}
 }