+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
 
 ### 🗑️ Deprecation
-- [](https://github.com/vegaprotocol/vega/issues/xxxx) -
+- [5513](https://github.com/vegaprotocol/vega/issues/5513) - Remove all checkpoint restore command
 
 ### 🛠 Improvements
 - [5428](https://github.com/vegaprotocol/vega/pull/5428) - Update contributor information
@@ -19,6 +19,7 @@
 - [5508](https://github.com/vegaprotocol/vega/issues/5508) - Fix duplicated staking events
 - [5514](https://github.com/vegaprotocol/vega/issues/5514) - Emit `rewardScore` event correctly after loading from checkpoint
 - [5520](https://github.com/vegaprotocol/vega/issues/5520) - Do not fail silently when wallet fails to start
+- [5521](https://github.com/vegaprotocol/vega/issues/5521) - Fix asset bundle and add asset status
 
 ## 0.52.0